+++ src/Console/Filesystem/FilesystemMvCommand.php
@@ -1,0 +1,95 @@
+<?php
+
+namespace ConductorCore\Console\Filesystem;
+
+use ConductorCore\Filesystem\MountManager\MountManager;
+use ConductorCore\MonologConsoleHandlerAwareTrait;
+use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+class FilesystemMvCommand extends Command
+{
+    use MonologConsoleHandlerAwareTrait;
+
+    /**
+     * @var MountManager
+     */
+    private $mountManager;
+    /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
+    /**
+     * DatabaseExportCommand constructor.
+     *
+     * @param MountManager         $mountManager
+     * @param LoggerInterface|null $logger
+     * @param string|null          $name
+     */
+    public function __construct(
+        MountManager $mountManager,
+        LoggerInterface $logger = null,
+        string $name = null
+    ) {
+        $this->mountManager = $mountManager;
+        if (is_null($logger)) {
+            $logger = new NullLogger();
+        }
+        $this->logger = $logger;
+        parent::__construct($name);
+    }
+
+    /**
+     * @return void
+     */
+    protected function configure()
+    {
+        $filesystemAdapterNames = $this->mountManager->getFilesystemPrefixes();
+        $this->setName('filesystem:mv')
+            ->addArgument(
+                'source',
+                InputArgument::REQUIRED,
+                "Source path in the format {adapter}://{path}.\nAvailable Adapters: <comment>" . implode(
+                    ', ',
+                    $filesystemAdapterNames
+                )
+                . '</comment>'
+            )
+            ->addArgument(
+                'destination',
+                InputArgument::REQUIRED,
+                "Destination path in the format {adapter}://{path}.\nAvailable Adapters: <comment>" . implode(
+                    ', ',
+                    $filesystemAdapterNames
+                )
+                . '</comment>'
+            )
+            ->setDescription(
+                'Move a file or directory from a source filesystem to a destination filesystem.'
+            )
+            ->setHelp(
+                "This command moves a file or directory from a source filesystem to a destination filesystem."
+            );
+    }
+
+    /**
+     * @param InputInterface  $input
+     * @param OutputInterface $output
+     *
+     * @return int
+     */
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $this->injectOutputIntoLogger($output, $this->logger);
+        $source = $input->getArgument('source');
+        $destination = $input->getArgument('destination');
+        $this->mountManager->move($source, $destination);
+        return 0;
+    }
+
+}